Work split functions. Fzf defaults. For loop snip.
My <Leader>f mapping to open a 4 way split in vim didn't work... so put
it in a function and call the function (using <Leader>f).

Create a bash function for dapony (tms) to make a 4 way tmux split and
resize the panes for how I usually work.

Make fzf select multiple file be default.

Fix matlab for loop snippet so it allows user to input end of range
($3) e.g. for $1 = $2:$3