Create commands to toggle light colorscheme
I made a file to be source by tmux (tmux_light) that uses light
colorschemes. Using <Leader>o in vim will toggle both the colorscheme in
vim and the colorscheme in the tmux terminal sessions.

The prompt is now a color that works in both light and dark colorschemes.

I also made aliases: ol (On-Light) and od (Off-Dark)