Fix a bug with the vim-tmux color switching
If I switched vim to light, then created a new tmux split, then switched
back to vim, vim would toggle to dark... this was because the tmux.conf
would run in the new tmux split, and would first launch tmux, using the
light theme, not yet have a THEME varible defined for itself, set that
to dark - then vim picked up the dark theme.

I now check to see if the tmux environment has been set to THEME=light,
and only if not do I set it to THEME=dark. This cleared up the issue.