Fix refactor function. Add mapping.
Use the pattern '^import | from ' to find the last import line (instead
of just the pattern '^import '.

Add mapping to call a function which puts a blank line above and below
the current line. Benefit of letting a function handle it means I can
give a count to make a nice big space around a line.